All the parsers are working with the commands for the basics

+++ src/slogo/model/backEndInternal/CommandExecutor.java
@@ -4,6 +4,8 @@ import slogo.model.backEndInternal.commands.Command;
 import slogo.model.ExecutionException;
 import slogo.model.Executor;
 
+import java.lang.reflect.Method;
+
 public class CommandExecutor implements Executor {
 
     public CommandExecutor(){
@@ -15,8 +17,19 @@ public class CommandExecutor implements Executor {
      */
     public Object executeCommand(Command currentCommand) throws ExecutionException {
        //System.out.println("So far good"+currentCommand.execute().toString());
+//        //currentCommand.getClass().getMethods()[0].getReturnType();
+//        try {
+//            Method testMethod = currentCommand.getClass().getMethod("execute");
+//           if(testMethod.getReturnType().equals(Integer.TYPE)){
+//               return (Integer)currentCommand.execute();
+//            }
+//
+//            //testMethod.getReturnType()
+//        } catch (NoSuchMethodException e) {
+//            e.printStackTrace();
+//        }
 
-       return (Double) currentCommand.execute();
+        return currentCommand.execute();
     }
