+++ src/slogo/model/backEndInternal/commands/And.java
@@ -11,7 +11,7 @@ public class And implements Command<Integer> {
   }
 
   @Override
-  public Integer execute() {
+  public double execute() {
     return (value1 != 0 && value2 != 0 ) ? 1 : 0;
   }
 }
