+++ src/slogo/model/backEndInternal/commands/PenUp.java
@@ -1,7 +1,6 @@
 package slogo.model.backEndInternal.commands;
 
 import slogo.model.backEndInternal.BackEndTurtle;
-import slogo.model.backEndInternal.commands.Command;
 
 public class PenUp implements Command<Integer> {
 
@@ -12,7 +11,7 @@ public class PenUp implements Command<Integer> {
   }
 
   @Override
-  public Integer execute() {
+  public double execute() {
     backEndTurtle.flipPen();
     return 0;
   }
