+++ src/slogo/model/backEndInternal/commands/SetTowards.java
@@ -1,22 +1,21 @@
 package slogo.model.backEndInternal.commands;
 
 import slogo.model.backEndInternal.BackEndTurtle;
-import slogo.model.backEndInternal.commands.Command;
 
 public class SetTowards implements Command<Double> {
 
   private BackEndTurtle backEndTurtle;
-  private int x;
-  private int y;
+  private double x;
+  private double y;
 
- public SetTowards(BackEndTurtle myTurtle, int val1, int val2) {
+ public SetTowards(BackEndTurtle myTurtle, double val1, double val2) {
     this.backEndTurtle = myTurtle;
     this.x = val1;
     this.y = val2;
   }
 
   @Override
-  public Double execute() {
+  public double execute() {
     double newHeading = Math.toDegrees(Math.atan2(y, x));
     backEndTurtle.setHeading(newHeading);
     return newHeading;
